deletes old dropdown made by yasharth
