+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@aftershootco/pegasus",
-	"version": "1.0.27",
+	"version": "1.0.28",
 	"description": "The core UI library for Aftershoot",
 	"author": "Aftershoot",
 	"license": "ISC",
